Add lua api for string data mapper
